+++ be/src/main/java/com/foodymoody/be/feed/repository/FeedRepository.java
@@ -1,7 +1,7 @@
 package com.foodymoody.be.feed.repository;
 
 import com.foodymoody.be.feed.domain.Feed;
-import com.foodymoody.be.member.repository.MemberProfileFeedPreviewResponse;
+import com.foodymoody.be.feed.repository.dto.MemberProfileFeedPreviewResponse;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 public interface FeedRepository extends JpaRepository<Feed, String> {
-    @Query(value = "SELECT new com.foodymoody.be.member.repository.MemberProfileFeedPreviewResponse (f.id, i.url) FROM Feed f "
+    @Query(value = "SELECT new com.foodymoody.be.feed.repository.dto.MemberProfileFeedPreviewResponse (f.id, i.url) FROM Feed f "
             + "INNER JOIN f.imageMenus.imageMenusList im "
             + "INNER JOIN FETCH Image i ON im.imageId = i.id "
             + "WHERE f.memberId = :memberId AND im.displayOrder = 0 "
