+++ be/src/main/java/com/foodymoody/be/member/controller/MemberController.java
@@ -1,6 +1,6 @@
 package com.foodymoody.be.member.controller;
 
-import com.foodymoody.be.member.repository.MemberProfileFeedPreviewResponse;
+import com.foodymoody.be.feed.repository.dto.MemberProfileFeedPreviewResponse;
 import com.foodymoody.be.member.repository.MemberProfileResponse;
 import com.foodymoody.be.member.controller.dto.TasteMoodResponse;
 import com.foodymoody.be.member.controller.dto.MemberSignupRequest;
